revert centroid changes
